+++ lib/core/commentable.dart
@@ -1,10 +1,8 @@
 import '../models/comment.dart';
 import 'authored.dart';
-import 'item.dart';
 import 'paginated_item.dart';
 
-abstract class CommentableItem
-    implements PaginatedItem<Comment>, Item, Authored {
+abstract class CommentableItem implements PaginatedItem<Comment>, Authored {
   Future<bool> subscribe();
   Future<bool> unsubscribe();
   bool get canComment;