+++ lib/models/event_attendees.dart
@@ -1,142 +1,122 @@
-import 'package:cached_network_image/cached_network_image.dart';
 import 'package:flutter/material.dart';
 
 import '../constants.dart';
 import '../services/microcosm_client.dart';
 import '../core/paginated.dart';
-import '../widgets/profile_sheet.dart';
-import 'profile.dart';
+import '../widgets/attendee_chip.dart';
+import '../widgets/attendee_shimmer.dart';
+import 'attendee.dart';
 
-// TODO: refactor this whole file
-// Maybe make it an ItemWithChildren
-
-class EventAttendees implements Paginated<Profile> {
+class EventAttendees implements Paginated<Attendee> {
   final int eventId;
-  final int attendeeCount;
-  final int pageSize = 250;
-  List<Profile>? attendees;
+  @override
+  final int startPage;
+  @override
+  int totalChildren;
 
-  EventAttendees({
+  final Map<int, Attendee> _children = {};
+
+  EventAttendees.fromJson({
     required this.eventId,
-    required this.attendeeCount,
-  });
+    required Json json,
+    this.startPage = 0,
+  }) : totalChildren = json["attendees"]["total"] {
+    if (json.containsKey("items")) {
+      parsePage(json);
+    }
+  }
 
-  Future<List<Profile>> getPageOfChildren(int pageId) async {
-    Uri uri = Uri.parse(
-      "https://$API_HOST/api/v1/events/$eventId/attendees?limit=$pageSize&offset=${pageSize * pageId}",
+  @override
+  void parsePage(Json json) {
+    totalChildren = json["attendees"]["total"];
+
+    List<Attendee> attendees = json["attendees"]["items"]
+        .map<Attendee>(
+          (attendee) => Attendee.fromJson(json: attendee),
+        )
+        .toList();
+
+    for (final (index, attendee) in attendees.indexed) {
+      _children[json["attendees"]["offset"] + index] = attendee;
+    }
+  }
+
+  static Future<EventAttendees> getByEventId(int eventId) async {
+    Uri uri = Uri.https(
+      API_HOST,
+      "/api/v1/events/$eventId/attendees",
+      {
+        "limit": PAGE_SIZE.toString(),
+      },
     );
 
     Json json = await MicrocosmClient().getJson(uri);
 
-    List<Profile> items = json["attendees"]["items"]
-        .map<Profile>(
-          (item) => Profile.fromJson(
-            json: item["profile"],
-          ),
-        )
-        .toList();
-
-    return items;
+    return EventAttendees.fromJson(
+      eventId: eventId,
+      json: json,
+      startPage: json["attendees"]["page"] - 1,
+    );
   }
 
-  Future<List<Profile>> getAttendeeList() async {
-    attendees ??= await getPageOfChildren(0);
-    return attendees!;
+  @override
+  Future<void> loadPage(int pageId, {bool force = false}) async {
+    Uri uri = Uri.https(
+      API_HOST,
+      "/api/v1/events/$eventId/attendees",
+      {
+        "limit": PAGE_SIZE.toString(),
+        "offset": (PAGE_SIZE * pageId).toString(),
+      },
+    );
+
+    final bool lastPage = pageId == totalChildren ~/ PAGE_SIZE;
+    final int ttl = lastPage ? 5 : 3600;
+
+    Json json = await MicrocosmClient().getJson(
+      uri,
+      ttl: ttl,
+      ignoreCache: force,
+    );
+    parsePage(json);
   }
 
-  Widget getProfile(int index) {
-    return FutureBuilder(
-      future: getAttendeeList(),
+  @override
+  Future<void> resetChildren({bool force = false, int? childId}) async {
+    // With an attendee list we're not just appending new content,
+    // we're modifying the items on the list. So we can't just
+    // proactively request empty pages to see if there's new content.
+    final int pageId = (totalChildren - 1) ~/ PAGE_SIZE;
+
+    await loadPage(pageId, force: force);
+  }
+
+  @override
+  Widget childTile(int i) {
+    if (_children.containsKey(i)) {
+      return AttendeeChip(attendee: _children[i]!);
+    }
+    return FutureBuilder<Attendee>(
+      future: getChild(i),
       builder: (context, snapshot) {
         if (snapshot.hasData) {
-          // return const SizedBox(width: 100.0, height: 158.0);
-          Profile profile = snapshot.data![index];
-          return ActionChip(
-            visualDensity: VisualDensity.compact,
-            onPressed: () => showModalBottomSheet<void>(
-              enableDrag: true,
-              showDragHandle: true,
-              context: context,
-              builder: (BuildContext context) => ProfileSheet(
-                profile: profile,
-              ),
-            ),
-            avatar: CircleAvatar(
-              foregroundImage: CachedNetworkImageProvider(profile.avatar),
-              backgroundColor: Colors.grey.shade800,
-            ),
-            label: Text(profile.profileName),
-          );
+          return AttendeeChip(attendee: snapshot.data!);
         } else if (snapshot.hasError) {
-          return Icon(
-            Icons.error_outline,
-            color: Theme.of(context).colorScheme.error,
-            size: 24.0,
-          );
+          // TODO: something that indicates error
+          return const AttendeeShimmer();
         } else {
-          return const Center(
-            child: CircularProgressIndicator(),
-          );
+          return const AttendeeShimmer();
         }
       },
     );
   }
 
-  Widget build() {
-    List<Widget> profiles = [];
-
-    for (int i = 0; i < attendeeCount; i++) {
-      profiles.add(getProfile(i));
+  @override
+  Future<Attendee> getChild(int i) async {
+    if (_children.containsKey(i)) {
+      return _children[i]!;
     }
-
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.all(8.0),
-          child: Text("Attendees ($attendeeCount)"),
-        ),
-        Wrap(
-          spacing: 8.0, // gap between adjacent chips
-          runSpacing: 4.0, //
-          children: profiles,
-        )
-      ],
-    );
+    await loadPage(i ~/ PAGE_SIZE);
+    return _children[i]!;
   }
-
-  @override
-  Widget childTile(int i) {
-    // TODO: implement childTile
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<Profile> getChild(int i) {
-    // TODO: implement getChild
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<void> loadPage(int pageId, {bool force = false}) {
-    // TODO: implement loadOfChildren
-    throw UnimplementedError();
-  }
-
-  @override
-  void parsePage(Json json) {
-    // TODO: implement parsePage
-  }
-
-  @override
-  Future<void> resetChildren({bool force = false, int? childId}) async {
-    // TODO: implement resetChildren
-    throw UnimplementedError();
-  }
-
-  @override
-  // TODO: implement startPage
-  int get startPage => throw UnimplementedError();
-
-  @override
-  int get totalChildren => attendeeCount;
 }