+++ lib/widgets/event_header.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
+import 'package:provider/provider.dart';
 
+import '../models/attendee.dart';
 import '../models/event.dart';
+import '../models/user_provider.dart';
+import '../widgets/attendees_list.dart';
 
-class EventHeader extends StatelessWidget {
+class EventHeader extends StatefulWidget {
   const EventHeader({
     super.key,
     required this.event,
@@ -11,8 +15,15 @@ class EventHeader extends StatelessWidget {
 
   final Event event;
 
+  @override
+  State<EventHeader> createState() => _EventHeaderState();
+}
+
+class _EventHeaderState extends State<EventHeader> {
+  bool _updatingAttendance = false;
+
   List<Widget> _timeBlock(BuildContext context) {
-    if (event.when == null) {
+    if (widget.event.when == null) {
       return [
         const ListTile(
           leading: Icon(Icons.timer_outlined),
@@ -20,36 +31,39 @@ class EventHeader extends StatelessWidget {
         )
       ];
     }
-    if (event.multiDay == true) {
+
+    if (widget.event.multiDay == true) {
       return [
         ListTile(
           leading: const Icon(Icons.play_arrow_outlined),
           title: Text(
-            DateFormat('EEE d MMM, y h:mm a').format(event.start!),
+            DateFormat('EEE d MMM, y h:mm a').format(widget.event.start!),
           ),
           subtitle: _tzWarning(context),
         ),
         ListTile(
           leading: const Icon(Icons.stop_outlined),
           title: Text(
-            DateFormat('EEE d MMM, y h:mm a').format(event.end!),
+            DateFormat('EEE d MMM, y h:mm a').format(widget.event.end!),
           ),
+          subtitle: _expiryWarning(),
         ),
       ];
     }
 
-    if (event.multiDay == false) {
+    if (widget.event.multiDay == false) {
       return [
         ListTile(
           leading: const Icon(Icons.calendar_month),
           title: Text(
-            DateFormat('EEE d MMM, y').format(event.start!),
+            DateFormat('EEE d MMM, y').format(widget.event.start!),
           ),
+          subtitle: _expiryWarning(),
         ),
         ListTile(
           leading: const Icon(Icons.watch_later_outlined),
           title: Text(
-            "From ${DateFormat.jm().format(event.start!)} to ${DateFormat.jm().format(event.end!)}",
+            "From ${DateFormat.jm().format(widget.event.start!)} to ${DateFormat.jm().format(widget.event.end!)}",
           ),
           subtitle: _tzWarning(context),
         ),
@@ -59,38 +73,98 @@ class EventHeader extends StatelessWidget {
     return [];
   }
 
-  Widget? _tzWarning(BuildContext context) {
-    if (event.equivalentTz() == false) {
-      return Text(
-        "⚠️ This event is in a different timezone",
-        style: TextStyle(
-          color: Theme.of(context).colorScheme.error,
-        ),
-      );
-    } else {
-      return const SizedBox.shrink();
-    }
-  }
+  Widget _tzWarning(BuildContext context) =>
+      widget.event.equivalentTz() == false
+          ? Text(
+              "⚠️ This event is in a different timezone",
+              style: TextStyle(
+                color: Theme.of(context).colorScheme.error,
+              ),
+            )
+          : const SizedBox.shrink();
+
+  Widget _expiryWarning() => widget.event.timingStatus == EventTiming.expired
+      ? const Text("This event has expired")
+      : const SizedBox.shrink();
 
   @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        ..._timeBlock(context),
-        ListTile(
-          leading: const Icon(Icons.location_on),
-          title: event.where != null
-              ? Text(event.where!)
-              : const Text(
-                  "TBD",
-                  style: TextStyle(
-                    fontStyle: FontStyle.italic,
-                    color: Colors.grey,
+  Widget build(BuildContext context) => Column(
+        children: [
+          ..._timeBlock(context),
+          ListTile(
+            leading: const Icon(Icons.location_on),
+            title: widget.event.where != null
+                ? Text(widget.event.where!)
+                : const Text(
+                    "TBD",
+                    style: TextStyle(
+                      fontStyle: FontStyle.italic,
+                      color: Colors.grey,
+                    ),
                   ),
-                ),
-        ),
-        event.getAttendees(),
-      ],
-    );
+          ),
+          AttendeesList(
+            futureAttendees: widget.event.getAttendees(),
+            initialAttendeeCount: widget.event.rsvpAttend,
+            rsvpAttend: widget.event.rsvpAttend,
+          ),
+          ..._attendanceButton(),
+        ],
+      );
+
+  List<Widget> _attendanceButton() {
+    int? userId = context.watch<UserProvider>().user?.id;
+    bool hasUser = userId != null;
+
+    if (hasUser && widget.event.timingStatus != EventTiming.expired) {
+      if (_updatingAttendance) {
+        return [
+          ElevatedButton.icon(
+            icon: const SizedBox(
+              width: 18.0,
+              height: 18.0,
+              child: CircularProgressIndicator(),
+            ),
+            label: const Text("Loading"),
+            onPressed: null,
+          ),
+        ];
+      }
+      if (widget.event.flags.attending) {
+        return [
+          ElevatedButton.icon(
+            onPressed: () async {
+              setState(() => _updatingAttendance = true);
+              await widget.event.updateAttendance(
+                userId,
+                AttendeeStatus.no,
+              );
+              await widget.event.resetChildren();
+              if (mounted) setState(() => _updatingAttendance = false);
+            },
+            icon: const Icon(Icons.close_outlined),
+            label: const Text("Flounce"),
+          ),
+        ];
+      } else {
+        return [
+          ElevatedButton.icon(
+            onPressed: () async {
+              setState(() => _updatingAttendance = true);
+              await widget.event.updateAttendance(
+                userId,
+                AttendeeStatus.yes,
+              );
+              await widget.event.resetChildren();
+              if (mounted) setState(() => _updatingAttendance = false);
+            },
+            icon: const Icon(Icons.add),
+            label: const Text("Attend"),
+          ),
+        ];
+      }
+    } else {
+      return [];
+    }
   }
 }